Renamed `Lists` folder to `OptionsList`, as well as `ListMessage` to `OptionsListMessage` and `RequestListMessage` to `RequestOptionsListMessage`.
